#11340: migrate move to ttnn and delete remnants of ttlib copy (#11343)
