+++ tests/tt_eager/python_api_testing/trace_testing/misc/test_bert_ops.py
@@ -136,7 +136,7 @@ class TestBertOpsTrace:
                     ttnn.ShardOrientation.COL_MAJOR,
                 )
             else:
-                in0_t = ttnn.experimental.tensor.clone(in0_t_res, interleaved_mem_config_L1)
+                in0_t = ttnn.clone(in0_t_res, memory_config=interleaved_mem_config_L1)
 
             if has_bias:
                 output_t = ttnn.linear(
