+++ ttnn/cpp/ttnn/operations/data_movement/move/move_pybind.hpp
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "pybind11/pybind_fwd.hpp"
+
+namespace ttnn::operations::data_movement::detail {
+
+void py_bind_move(pybind11::module& m);
+
+}  // namespace ttnn::operations::data_movement::detail
