+++ ttnn/cpp/ttnn/operations/data_movement/permute/permute.cpp
@@ -6,7 +6,6 @@
 
 #include "ttnn/common/constants.hpp"
 #include "ttnn/operations/data_movement/transpose/transpose.hpp"
-#include "ttnn/deprecated/tt_dnn/op_library/copy/copy_op.hpp"
 
 #include "tt_metal/host_api.hpp"
 #include "tt_metal/common/constants.hpp"
